+++ src/components/index.js
@@ -74,3 +74,4 @@ export { default as Clickable } from "./Clickable/Clickable";
 export { default as ColorUtils } from "../utils/colors-utils";
 export { default as Slider } from "./Slider/Slider";
 export { default as IconButton } from "./IconButton/IconButton";
+export { default as Flex } from "./Flex/Flex";