+++ src/constants/positions.js
@@ -1,0 +1,5 @@
+export const BASE_POSITIONS = {
+  START: "Start",
+  CENTER: "Center",
+  END: "End"
+};