+++ src/constants/sizes.js
@@ -1,13 +1,11 @@
-export const SIZES_BASIC = Object.freeze({
-  SMALL: "small",
-  MEDIUM: "medium",
-  LARGE: "large"
-});
+export const PASCAL_BASE_SIZE = Object.freeze({ SMALL: "Small", MEDIUM: "Medium", LARGE: "Large" });
+
+export const BASE_SIZES = Object.freeze({ SMALL: "small", MEDIUM: "medium", LARGE: "large" });
 
 export const SIZES = Object.freeze({
   XXS: "xxs",
   XS: "xs",
-  ...SIZES_BASIC
+  ...BASE_SIZES
 });
 
 export const DialogPositions = Object.freeze({